+++ spring-cloud-tencent-plugin-starters/spring-cloud-starter-tencent-mq-plugin/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.tencent.cloud.plugin.lane.kafka.KafkaLaneAspectConfiguration
+  com.tencent.cloud.plugin.mq.lane.kafka.KafkaLaneAspectConfiguration
