+++ spring-cloud-tencent-plugin-starters/spring-cloud-starter-tencent-mq-plugin/src/test/java/com/tencent/cloud/plugin/mq/lane/tsf/TsfActiveLaneTest.java
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package com.tencent.cloud.plugin.lane.tsf;
+package com.tencent.cloud.plugin.mq.lane.tsf;
 
 import java.util.Arrays;
 import java.util.Collections;
